Changed time to long and some cosmetic changes.

+++ diffusion_help.h
@@ -16,13 +16,13 @@ typedef __m128 grid_simd_type;
 #	endif /* DOUBLE */
 #endif /* NO_SSE */
 
-#	ifdef DOUBLE
-#		define MPI_GRID_TYPE MPI_DOUBLE
+#ifdef DOUBLE
+#	define MPI_GRID_TYPE MPI_DOUBLE
 typedef double grid_type;
-#	else
-#		define MPI_GRID_TYPE MPI_FLOAT
+#else
+#	define MPI_GRID_TYPE MPI_FLOAT
 typedef float grid_type;
-#	endif /* DOUBLE */
+#endif /* DOUBLE */
 
 #define FIRST_FREE_TAG 9
 
@@ -41,11 +41,11 @@ enum { X_COORD,
 void send_grid(grid_type **grid, size_t *grains, int *offset, int rank,
 		MPI_Comm comm, double *time_comm, int base_tag);
 
-void recv_grid(grid_type **grid, size_t *grains, int *offset, int time,
+void recv_grid(grid_type **grid, size_t *grains, int *offset, long time,
 		int rank, int nnodes, MPI_Comm comm, double *time_comm, int base_tag, 
-		void (*handler)(int, int, int, size_t*, int*, grid_type*, void *), void *handlerargs);
+		void (*handler)(long, int, int, size_t*, int*, grid_type*, void *), void *handlerargs);
 	
-void print_elem(int time, int rank, int x, size_t *grains, int *offset, grid_type *column,
+void print_elem(long time, int rank, int x, size_t *grains, int *offset, grid_type *column,
 		void *fd);
 
 #endif
